New "Customize / Rest Code Generation Templates".
By default, Xrm Tools will use default templates and templates will not be generated under solution structure. This command allows the developer to generated them only when they want to customize them. This can be done at solution or project level.

The command also allows the developer to reset an individual template, all templates of a project or templates at the solution level.