Special treatment of `Config` and `SecureConfig`
When a string parameter in the constructor of a dependency has a name that ends with "config" or "secureconfig" regardless of casing, DI will try to pass the value of Config or SecureConfig property. The respective property must exist in the plugin's type (or inherited).

It's possible for the parameter to have a different name if `[Dependency("Config")]` or `[Dependency("SecureConfig")]` is applied on the parameter.